requirements.txt
Flask>=2.3.0
Flask-SocketIO>=5.3.0
python-dotenv>=1.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Caching>=2.1.0
redis>=5.0.0
requests>=2.31.0
google-genai>=0.7.0
google-generativeai>=0.8.0
python-docx>=1.1.0
odfpy>=1.4.0
psycopg2-binary>=2.9.7
PyPDF2>=3.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
gunicorn>=21.2.0
sqlalchemy>=2.0.0